[FEAT] update get citizen profile

+++ suwmp_be/src/main/java/com/example/suwmp_be/dto/citizen_profile/CitizenProfileGetResponse.java
@@ -10,7 +10,7 @@ public record CitizenProfileGetResponse(
         String email,
         Instant createdAt,
         int points,
-        int reports,
+        long reports,
         double volume,
-        int feedbacks
+        long feedbacks
 ) { }
